InternetandCloud/BubbleTeaStore(Web app)/requirements.txt
flask
google-cloud-datastore
gunicorn

